Added new component ChessOpeningSearchResult
