+++ src/components/Dialog/ChessOpeningSearchEcoDialog.js
@@ -1,15 +1,11 @@
 import React, { useState } from 'react';
 import { useDispatch, useSelector } from "react-redux";
 import { makeStyles } from '@mui/styles';
-import { Button, Dialog, DialogActions, DialogContent, DialogTitle, IconButton,
-  MenuItem, Paper, Table, TableBody, TableCell, TableContainer, TableRow, TextField
-} from '@mui/material';
+import { Button, Dialog, DialogActions, DialogContent, DialogTitle, MenuItem, TextField } from '@mui/material';
 import ChessOpeningSearchAjaxLoader from "../AjaxLoader/ChessOpeningSearchAjaxLoader.js";
-import PublishIcon from '@mui/icons-material/Publish';
-import { wsMssgStartLoadpgn, wsMssgQuit } from "../../actions/serverActions";
+import ChessOpeningSearchResult from "./ChessOpeningSearchResult.js";
 import chessOpeningSearchAjaxLoaderActionTypes from '../../constants/ajaxLoader/chessOpeningSearchAjaxLoaderActionTypes';
 import chessOpeningSearchEcoDialogActionTypes from '../../constants/dialog/chessOpeningSearchEcoDialogActionTypes';
-
 
 const useStyles = makeStyles({
   form: {
@@ -22,14 +18,6 @@ const ChessOpeningSearchEcoDialog = () => {
   const state = useSelector(state => state);
   const [openings, setOpenings] = useState([]);
   const dispatch = useDispatch();
-
-  const handleLoad = (movetext) => {
-    wsMssgQuit(state).then(() => {
-      wsMssgStartLoadpgn(state, movetext).then(() => {
-        dispatch({ type: chessOpeningSearchEcoDialogActionTypes.CLOSE });
-      });
-    });
-  };
 
   const handleChange = (event) => {
     event.preventDefault();
@@ -88,30 +76,7 @@ const ChessOpeningSearchEcoDialog = () => {
           </DialogActions>
         </form>
         <ChessOpeningSearchAjaxLoader />
-        <TableContainer component={Paper}>
-          <Table stickyHeader aria-label="simple table">
-            <TableBody>
-              {
-                openings.map((item, i) => (
-                  <TableRow key={i}>
-                    <TableCell align="right">{item.eco}</TableCell>
-                    <TableCell align="right">{item.name}</TableCell>
-                    <TableCell align="right">{item.movetext}</TableCell>
-                    <TableCell align="right">
-                      <IconButton
-                        aria-label="load"
-                        color="primary"
-                        onClick={() => handleLoad(item.movetext)}
-                      >
-                        <PublishIcon />
-                      </IconButton>
-                    </TableCell>
-                  </TableRow>
-                ))
-              }
-            </TableBody>
-          </Table>
-        </TableContainer>
+        <ChessOpeningSearchResult props={{ openings: openings }} />
       </DialogContent>
     </Dialog>
   );
